Move editor tests into existing test case.
